+++ Src/Pscx/Commands/EnvironmentBlock/AddPathVariableCommand.cs
@@ -1,17 +1,17 @@
 // Copyright © 2026 PowerShell Core Community Extensions Team. All rights reserved.
 // Licensed under MIT license.
 
-using Pscx.EnvironmentBlock;
 using System.ComponentModel;
 using System.Management.Automation;
 
 namespace Pscx.Commands.EnvironmentBlock {
-    [Cmdlet(VerbsCommon.Add, PscxNouns.PathVariable), Description("Adds values to an environment variable of type PATH (default is PATH variable)")]
+    [OutputType(typeof(PathVariableChange))]
+    [Cmdlet(VerbsCommon.Add, PscxNouns.PathVariable, SupportsShouldProcess = true), Description("Adds values to an environment variable of type PATH (default is PATH variable)")]
     [RelatedLink(typeof(GetPathVariableCommand))]
     [RelatedLink(typeof(SetPathVariableCommand))]
     [RelatedLink(typeof(PopEnvironmentBlockCommand))]
     [RelatedLink(typeof(PushEnvironmentBlockCommand))]
-    public sealed class AddPathVariableCommand : PathVariableCommandBase {
+    public sealed class AddPathVariableCommand : PathVariableMutationCommandBase {
         [Parameter(Position = 0, Mandatory = true, ValueFromPipeline = true)]
         public string[] Value { get; set; }
 
@@ -21,13 +21,18 @@ namespace Pscx.Commands.EnvironmentBlock {
         [Parameter]
         public SwitchParameter Prepend { get; set; }
 
+        protected override void ProcessRecord() {
+            CollectValues(Value);
+        }
+
         protected override void EndProcessing() {
-            using PathVariable variable = new(Name, Target);
-            if (Prepend) {
-                variable.Prepend(Value);
-            } else {
-                variable.Append(Value);
-            }
+            ApplyChange("Add", variable => {
+                if (Prepend) {
+                    variable.Prepend(CollectedValues);
+                } else {
+                    variable.Append(CollectedValues);
+                }
+            });
         }
     }
 }