+++ Src/Pscx/Commands/EnvironmentBlock/RemovePathVariableCommand.cs
@@ -1,36 +1,30 @@
 ﻿// Copyright © 2023 PowerShell Core Community Extensions Team. All rights reserved.
 // Licensed under MIT license.
 
-using Pscx.EnvironmentBlock;
-using System;
 using System.ComponentModel;
 using System.Management.Automation;
 
 namespace Pscx.Commands.EnvironmentBlock {
+    [OutputType(typeof(PathVariableChange))]
     [Cmdlet(VerbsCommon.Remove, PscxNouns.PathVariable, SupportsShouldProcess = true), Description("Removes values from an environment variable of type PATH (default is PATH variable)")]
     [RelatedLink(typeof(AddPathVariableCommand))]
     [RelatedLink(typeof(GetPathVariableCommand))]
     [RelatedLink(typeof(SetPathVariableCommand))]
     [RelatedLink(typeof(PopEnvironmentBlockCommand))]
     [RelatedLink(typeof(PushEnvironmentBlockCommand))]
-    public class RemovePathVariableCommand : PathVariableCommandBase {
+    public class RemovePathVariableCommand : PathVariableMutationCommandBase {
         [Parameter(Position = 0, Mandatory = true, ValueFromPipeline = true)]
         public string[] Value { get; set; }
 
         [Parameter]
         public override string Name { get; set; }
 
+        protected override void ProcessRecord() {
+            CollectValues(Value);
+        }
+
         protected override void EndProcessing() {
-            if ((Value?.Length ?? 0) == 0) {
-                return;
-            }
-
-            if (!ShouldProcess(Name, "Remove values from path-like environment variable")) {
-                return;
-            }
-
-            using PathVariable variable = new(Name, Target);
-            variable.Remove(Value);
+            ApplyChange("Remove", variable => variable.Remove(CollectedValues));
         }
     }